pyproject.toml
[tool.poetry]
name = "adk-test"
version = "0.1.0"
description = ""
authors = ["Corey Munn <michael.munn@secunetics.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
google-adk = "^1.7.0"
google-generativeai = "^0.8.5"
pydantic = "^2.11.7"
google-genai = "^1.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
